# sogapeint-CRM/sogapaint-crm/src/assets/swagger-docs/swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "API Sogapeint",
      "version": "1.0.0",
      "description": "API backend de l'application Sogapeint."
    },
    "host": "localhost:3000",
    "basePath": "/",
    "schemes": [
      "http"
    ],
    "paths": {
...
      "/kpi/ordersByDate": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "startDate",
              "in": "query",
              "type": "string"
            },
            {
              "name": "endDate",
              "in": "query",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
...
      "/kpi/activeOrdersByDate": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "startDate",
              "in": "query",
              "type": "string"
            },
            {
              "name": "endDate",
              "in": "query",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
...
      "/kpi/inactiveOrdersByDate": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "startDate",
              "in": "query",
              "type": "string"
            },
            {
              "name": "endDate",
              "in": "query",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
...
      "/kpi/completedOrdersByDate": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "startDate",
              "in": "query",
              "type": "string"
            },
            {
              "name": "endDate",
              "in": "query",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/kpi/averageExecutionTime": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "startDate",
              "in": "query",
              "type": "string"
            },
            {
              "name": "endDate",
              "in": "query",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/kpi/totalRevenue": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "startDate",
              "in": "query",
              "type": "string"
            },
            {
              "name": "endDate",
              "in": "query",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/kpi/averageRevenue": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "startDate",
              "in": "query",
              "type": "string"
            },
            {
              "name": "endDate",
              "in": "query",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
...
      "/kpi/ordersByServiceByDate": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "startDate",
              "in": "query",
              "type": "string"
            },
            {
              "name": "endDate",
              "in": "query",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/kpi/ordersWithIncidents": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "startDate",
              "in": "query",
              "type": "string"
            },
            {
              "name": "endDate",
              "in": "query",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/kpi/mailResponseRate": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "startDate",
              "in": "query",
              "type": "string"
            },
            {
              "name": "endDate",
              "in": "query",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
...
      "/kpi/averageBillingAmount": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "startDate",
              "in": "query",
              "type": "string"
            },
            {
              "name": "endDate",
              "in": "query",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
...
      "/kpi/ordersStatusDistribution": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "startDate",
              "in": "query",
              "type": "string"
            },
            {
              "name": "endDate",
              "in": "query",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/kpi/customerRenewalRate": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "startDate",
              "in": "query",
              "type": "string"
            },
            {
              "name": "endDate",
              "in": "query",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/kpi/order-amount": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "userId",
              "in": "query",
              "type": "string"
            },
            {
              "name": "dateStart",
              "in": "query",
              "type": "string"
            },
            {
              "name": "dateEnd",
              "in": "query",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/kpi/order-amount-this-month": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "userId",
              "in": "query",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/kpi/average-order-amount": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "userId",
              "in": "query",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
...
      "/ordersByDate": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "startDate",
              "in": "query",
              "type": "string"
            },
            {
              "name": "endDate",
              "in": "query",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
...
      "/activeOrdersByDate": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "startDate",
              "in": "query",
              "type": "string"
            },
            {
              "name": "endDate",
              "in": "query",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
...
      "/inactiveOrdersByDate": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "startDate",
              "in": "query",
              "type": "string"
            },
            {
              "name": "endDate",
              "in": "query",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
...
      "/completedOrdersByDate": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "startDate",
              "in": "query",
              "type": "string"
            },
            {
              "name": "endDate",
              "in": "query",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/averageExecutionTime": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "startDate",
              "in": "query",
              "type": "string"
            },
            {
              "name": "endDate",
              "in": "query",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/totalRevenue": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "startDate",
              "in": "query",
              "type": "string"
            },
            {
              "name": "endDate",
              "in": "query",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/averageRevenue": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "startDate",
              "in": "query",
              "type": "string"
            },
            {
              "name": "endDate",
              "in": "query",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
...
      "/ordersByServiceByDate": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "startDate",
              "in": "query",
              "type": "string"
            },
            {
              "name": "endDate",
              "in": "query",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/ordersWithIncidents": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "startDate",
              "in": "query",
              "type": "string"
            },
            {
              "name": "endDate",
              "in": "query",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/mailResponseRate": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "startDate",
              "in": "query",
              "type": "string"
            },
            {
              "name": "endDate",
              "in": "query",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
...
      "/averageBillingAmount": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "startDate",
              "in": "query",
              "type": "string"
            },
            {
              "name": "endDate",
              "in": "query",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
...
      "/ordersStatusDistribution": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "startDate",
              "in": "query",
              "type": "string"
            },
            {
              "name": "endDate",
              "in": "query",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/customerRenewalRate": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "startDate",
              "in": "query",
              "type": "string"
            },
            {
              "name": "endDate",
              "in": "query",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/order-amount": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "userId",
              "in": "query",
              "type": "string"
            },
            {
              "name": "dateStart",
              "in": "query",
              "type": "string"
            },
            {
              "name": "dateEnd",
              "in": "query",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/order-amount-this-month": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "userId",
              "in": "query",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/average-order-amount": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "userId",
              "in": "query",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
...
    }
  }